// src/app/model/SuspenseAccount.ts
export class SuspenseAccount {
  Id: number;
  DateOfCredit: Date;
  DateOfCreditStr: string;
  BankName: string;
  AmountCredited: number;
  CreditDetails: string;
  Status: number;
  LastUpdatedOn: Date;
  LastUpdatedBy: number;
  ClaimId: number;
  ConsultantId: number;
  InvoiceNumber: string;
  InvoiceDate: Date;
  ClientName: string;
  ConsultantName: string;
}