// src/app/model/SuspenseAccountClaimDetail.ts
export class SuspenseAccountClaimDetail {
  Id: number;
  SuspenseAccountID: number;
  ClientID: number;
  InvoiceID: number;
  DocumentID: number;
  Remarks: string;
  Status: number;
  LastUpdatedOn: Date;
  LastUpdatedBy: number;
  ConsultantId: number;
  ApprovedOn: Date;
  ApprovedBy: number;
  ApproverRemarks: string;
  FileName: string;
}